+++ structurizr-core/src/main/java/com/structurizr/view/Shape.java
@@ -16,6 +16,7 @@ public enum Shape {
     Folder,
     WebBrowser,
     Window,
+    Terminal,
     Shell,
     MobileDevicePortrait,
     MobileDeviceLandscape,